minicore: rename `core::arch` to `core::core_arch`
`core::arch` is actually `core::core_arch` in Rust's `core`.
